+++ main/Motor.hpp
@@ -48,12 +48,11 @@
 
 void motorInit(); //Initializes motors
 
-uint8_t getDesiredRPM(uint8_t index); //gets the correct RPM for either side
-
+int getDesiredRPM(uint8_t index); //gets the correct RPM for either side
 /* Wrapper functions for updateDesiredRPM */
 void updateDesiredRPMLeft(int change);
 void updateDesiredRPMRight(int change);
-
+/*                           */
 void updateDesiredRPM(uint8_t side, int change); //Updates the global desired speed for each motor (this happens through the controller from ESP-NOW), the user will be inputting a desired RPM, which the interface displays as a velocity but that would be converted in code
 void matchDesiredRPM(); //Updates each motors speed to be the global desired speed, can incorporate a PID in the future
